+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.1
+- Fix ConfigPool load cache bugs
+- Fix FileSource load() return value
+
 ## 0.1.0
 - Start project
 - Add FileSource